+++ R/new_project.R
@@ -2,10 +2,9 @@
 #'@param prefix character. 3 character, alphabetic uppercase prefix to be used for the project. Usually the first three letters of the organization name.
 #'@param year numeric. year the project was started in, e.g. 2019.
 #'@param month numeric. month the project was started in 
-#'@param data_folder character. path to data folder where the project folder should be created. starts at root of the project as defined by here::here. Defaults to "".
-#'@export
-new_project <- function(prefix, year, month, data_folder = "") {
-  
+#'@param data_folder character. path to data folder starting at root of the project. defaults to here::here()
+#'@export  
+new_project <- function(prefix, year, month, data_folder = here::here()) {
   usethis::ui_info(glue::glue("processing {prefix} {year} {month}"))
   # check validity of inputs
   if (!is.numeric(year) | nchar(as.character(year)) != 4) {
@@ -14,7 +13,7 @@ new_project <- function(prefix, year, month, data_folder = "") {
   if (!is.numeric(month) || month > 12 || month < 1) {
     usethis::ui_stop("Invalid month argument. It must be an integer between 1 and 12.")
   }
-  
+
   # check validity of prefix 
   if (!stringr::str_detect(prefix, "^[:upper:]{3,3}$")) usethis::ui_stop("Invalid prefix. It be 3 alphabetic characters long and it must be uppercase.")
 
@@ -25,8 +24,7 @@ new_project <- function(prefix, year, month, data_folder = "") {
   
   # project id path 
   project_id_path <- glue::glue("{year}-{month}-{prefix}")
-  dir.create(here::here(data_folder, project_id_path), showWarnings = FALSE)
-             
+  dir.create(fs::path(data_folder, project_id_path), showWarnings = FALSE)
   
   # meta data file (meta.json)
   template_meta <- get_meta_template()
@@ -37,7 +35,8 @@ new_project <- function(prefix, year, month, data_folder = "") {
   template_meta$year <- year
   template_meta$start <- glue::glue("{year}-{month}")
   
-  meta_path <- here::here(data_folder, project_id_path, "meta.json")
+  meta_path <- fs::path(data_folder, project_id_path, "meta.json")
+
   answer <- TRUE
   if (file.exists(meta_path)) {
     usethis::ui_warn("meta.json already exists in {meta_path}")
@@ -52,20 +51,37 @@ new_project <- function(prefix, year, month, data_folder = "") {
   # markdown files
   markdown_files <- c("00_about.md", "00_summary.md", "01_problem.md", "02_data.md", "03_approach.md", "04_impact.md")
   for (lang in c("de", "en")) {
-    dir.create(here::here(data_folder, project_id_path, lang), showWarnings = FALSE)
+    dir.create(fs::path(data_folder, project_id_path, lang), showWarnings = FALSE)
+
     purrr::walk(markdown_files, function(x) {
-        file_path <- here::here("data", project_id_path, lang, x)
+        file_path <- fs::path(data_folder, project_id_path, lang, x)
         
         answer <- TRUE
         if (file.exists(file_path)) {
           usethis::ui_warn(glue::glue("{file_path} already exists."))
           answer <- usethis::ui_yeah("Do you want to overwrite it?", yes = "Yes", no = "No", shuffle = FALSE)
         }
-        if (answer) file.create(here::here(data_folder, project_id_path, lang, x))
+        if (answer) file.create(file_path)
         usethis::ui_done(glue::glue("created {x} at {file_path}"))
       }
     )
   }
+}
+
+
+#'Use download applications script template
+#'@param project_id_path project id in path form, e.g. 2020-11-COR
+#'@param data_folder character. path to data folder starting at root of the project. defaults to "", i.e. root
+#'@export 
+use_fill_project <- function(project_id_path, data_folder = "") {
+  
+  usethis::use_template(
+    "fill_project.R",
+    save_as = fs::path(data_folder, project_id_path, "fill_project.R"),
+    data = list(project_id = project_id_path),
+    package = "projectutils",
+    open = TRUE
+  )
 }
 
 get_meta_template <- function() {