+++ tests/testthat/test-load-project.R
@@ -1,5 +1,0 @@
-context("loading projects")
-test_that("that non-project folders are ignored when loading projects", {
-  projects <- load_projects("tests/testthat/test_data/")
-  expect_equal(length(projects), 1)
-})